+++ lum-server/src/logger.js
@@ -14,19 +14,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // ============LICENSE_END=========================================================
+/* eslint-disable no-console */
 
 const winston = require("winston")
+    , path = require('path')
     , fs = require("fs")
     , utils = require('./utils');
 
-const logFolder = __dirname + '/../logs';
+const fileRotateSize = 100 * 1024 * 1024;
+
+const logFolder = path.join(__dirname, '../logs');
 try {if (!fs.existsSync(logFolder)) {fs.mkdirSync(logFolder);}}
 catch (e) {console.error(`failed to create log folder ${logFolder}`, e);}
-
+/**
+ * setup the log line format
+ */
 const logFormatText = winston.format.printf(({level, message, timestamp}) => {
   return `${timestamp} ${level.toUpperCase().padStart(10, ' ')}: ${message}`;
 });
-
+/**
+ * convert args to the logger into one line string
+ * @param  {function} original
+ */
 const logWrapper = (original) => {
     return (...args) => original(args.map(arg => {
             if (typeof arg === 'object') {return JSON.stringify(arg);}
@@ -36,37 +45,52 @@ const logWrapper = (original) => {
 };
 
 module.exports = {
-    initLogger(app_name) {
-        const transports = [
-            new (winston.transports.Console)()
-        ];
-        let logFile;
+    /**
+     * setup logger
+     * @param  {string} serverName name of the server
+     */
+    initLogger(serverName) {
+        const logFileName = (serverName || '') + '_' + new Date().toISOString().substr(0, 19).replace(/:/g, "") + ".log";
+
+        const logTo      = ['console'];
+        const transports = [new (winston.transports.Console)()];
         if (process.env.LOGDIR) {
-            logFile = logFolder + "/" + (app_name || '') + '_' + new Date().toISOString().substr(0, 19).replace(/:/g, "") + ".log";
-            transports.push(new (winston.transports.File)({ json: false, filename: logFile, maxsize: (100 * 1024 * 1024) }));
+            const logFile = path.join(logFolder, logFileName);
+            transports.push(new (winston.transports.File)({json: false, filename: logFile, maxsize: fileRotateSize}));
+            logTo.push(logFile);
         }
         const logger = winston.createLogger({
-            format: winston.format.combine(
-                winston.format.timestamp(),
-                winston.format.errors({stack: true}),
-                logFormatText
-            ),
+            format: winston.format.combine(winston.format.timestamp(), winston.format.errors({stack: true}),
+                                           logFormatText),
             transports: transports
         });
 
-        logger.error = logWrapper(logger.error);
-        logger.warn = logWrapper(logger.warn);
-        logger.info = logWrapper(logger.info);
+        logger.error   = logWrapper(logger.error);
+        logger.warn    = logWrapper(logger.warn);
+        logger.info    = logWrapper(logger.info);
         logger.verbose = logWrapper(logger.verbose);
-        logger.debug = logWrapper(logger.debug);
-        logger.silly = logWrapper(logger.silly);
+        logger.debug   = logWrapper(logger.debug);
+        logger.silly   = logWrapper(logger.silly);
 
         lumServer.logger = logger;
-        lumServer.logger.info("-----------------------------------------------------------------------");
 
-        lumServer.logger.info("logger started for", app_name, 'to', logFile? 'console': logFile);
-        if(process.env.NODE_ENV !== 'production'){
-            lumServer.logger.info('process.env', process.env);
+        lumServer.logger.info("-----------------------------------------------------------------------");
+        lumServer.logger.info("logger started for", serverName, 'to', logTo.join(' and '));
+
+        if (process.env.LOGDIR) {
+            const logFile = path.join(logFolder, 'healthcheck_' + logFileName);
+            const logForHealthcheck = winston.createLogger({
+                format: winston.format.combine(winston.format.timestamp(), winston.format.errors({stack: true}),
+                                               logFormatText),
+                transports: [new (winston.transports.File)({json: false, filename: logFile, maxsize: fileRotateSize})]
+            });
+            lumServer.logForHealthcheck = logWrapper(logForHealthcheck.info);
+            lumServer.logger.info("logForHealthcheck for", serverName, 'to', logFile);
+        } else {
+            lumServer.logForHealthcheck = () => {
+                /* no logging for healthcheck - too often and too many lines */
+            };
+            lumServer.logger.info("no logForHealthcheck for", serverName);
         }
     }
 };